Updates to Pipfile and api gateway refactor
