+++ Pipfile
@@ -19,4 +19,4 @@ pytest-cov = "*"
 keyring = "*"
 
 [requires]
-python_version = "3.6"
+python_version = "3.8"
